requirements.txt
PyMuPDF
python-docx
langchain-groq
langchain
fastapi
black